mesos-fetcher: do not access protobuf outside of libmesos
This fixes mesos-fetcher with protobuf v29.3. Linking statically against
protobuf from within libmesos (dynamically linked by mesos-fetcher) and
by mesos-fetcher directly causes conflicts in protobuf runtime.
