+++ src/launcher/do_fetcher.hpp
@@ -1,0 +1,3 @@
+
+int do_fetcher(int argc, char* argv[]);
+
